Fix alldocument
